Average Sales Per Unique Product Calculations
Avg Sales per Unique Product is an import metric for businesses to make strategic decisions. For example, if the sales per unique product are low, it may indicate that the product is not popular among customers & needs to be improved or discontinued. On the other hand, if the sales per unique product are high, it may indicate that the product is in high demand and the business should focus on producing more of it. This video shows how to compute the Avg Sales per Product. Watch Here: https://lnkd.in/euVrYcif

#PowerBI #DAX